manager/requirements.txt
psycopg2-binary
pytz
python-dotenv
